chore(tool): consolidate ex package tests into error_test.go
