+++ tool/ex/error_test.go
@@ -14,6 +14,46 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+func TestError(t *testing.T) {
+	err := Newf("a")
+	err = Wrapf(err, "b")
+	err = Wrap(Wrap(Wrap(err))) // make no sense
+	require.Contains(t, err.Error(), "a")
+	require.Contains(t, err.Error(), "b")
+
+	err = errors.New("c")
+	err = Wrapf(err, "d")
+	err = Wrapf(err, "e")
+	err = Wrap(Wrap(Wrap(err))) // make no sense
+	require.Contains(t, err.Error(), "c")
+	require.Contains(t, err.Error(), "d")
+}
+
+func TestJoinStackful(t *testing.T) {
+	e1 := New("first")
+	e2 := Newf("second %d", 2)
+	joined := Join(e1, e2)
+
+	require.ErrorIs(t, joined, e1)
+	require.ErrorIs(t, joined, e2)
+
+	var se *stackfulError
+	require.ErrorAs(t, joined, &se)
+}
+
+func TestJoinMixed(t *testing.T) {
+	stdErr := errors.New("std")
+	exErr := New("ex")
+	joined := Join(stdErr, exErr)
+
+	require.ErrorIs(t, joined, stdErr)
+	require.ErrorIs(t, joined, exErr)
+
+	var se *stackfulError
+	require.ErrorAs(t, joined, &se)
+	require.Contains(t, se.Error(), "ex")
+}
+
 // captureStderr redirects os.Stderr for the duration of fn and returns whatever
 // was written. printError writes directly to os.Stderr, so this lets us assert
 // on its output without spawning a subprocess.
